requirements.txt
Flask
Flask-Login
Flask-Bcrypt
pytesseract
PyMuPDF
pdf2image
Pillow
beautifulsoup4
requests
lxml
SQLAlchemy
python-Levenshtein
rapidfuzz
langchain
langchain-google-genai
python-dotenv
werkzeug
gunicorn